Define clean up steps and clean up code

+++ luaj-jse/src/main/java/org/luaj/vm2/ast/Str.java
@@ -10,7 +10,7 @@
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
-* 
+*
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -74,9 +74,9 @@ public class Str {
 				case '7':
 				case '8':
 				case '9':
-					int d = (int) (c[i++]-'0');
+					int d = c[i++]-'0';
 					for (int j = 0; i < n && j < 2 && c[i] >= '0' && c[i] <= '9'; i++, j++)
-						d = d*10+(int) (c[i]-'0');
+						d = d*10+c[i]-'0';
 					baos.write((byte) d);
 					--i;
 					continue;
